Move functions into modules to import and export
